This new playbook is to be run on the master and worker nodes. The dedicated files that I had before were pretty much the same except that the node playbook also ran the kube join command command.
I used VSCode with the yaml plugin to create this file, now the indents should look fine. Not only this but i put a line between all of the sub tasks so that it is much easier to read.

When it came to installing the kubernetes packages, I decided to remove the refresh apt because The first thing the playbook does is refresh the apt cache.